LayoutTests/platform/win/fast/forms/focus-selection-textarea-expected.txt

<textarea> focus selection

This test checks whether the selection is restored, cleared, or set to the full range when using different ways to focus a text area. These results all match Mozilla. When running manually, please follow the steps below. In the test harness, the test runs automatically.

1) Hit tab, only the word "SUCCESS" should be selected:
2) Hit Ctrl-Option-J (or Alt-J on Windows), only the word "SUCCESS" should be selected:
3) , only the word "SUCCESS" should be selected:
4) Click in the input field, nothing should be selected:
5) Click this label, only the word "SUCCESS" should be selected:
6) Hit Ctrl-Option-U (or Alt-U on Windows), only the word "SUCCESS" should be selected:
7) , only the word "SUCCESS" should be selected:
8)  , only the word "SUCCESS" should be selected:


9) Hit Ctrl-Option-I (or Alt-I on Windows), only the word "SUCCESS" should be selected:


PASS first.selectionStart is 11
PASS first.selectionEnd is 18
FAIL first.selectionStart should be 11. Was 0.
FAIL first.selectionEnd should be 18. Was 0.
PASS second.selectionStart is 11
PASS second.selectionEnd is 18
PASS second.selectionStart is 11
PASS second.selectionEnd is 18
PASS third.selectionStart is 11
PASS third.selectionEnd is 18
PASS third.selectionStart is 11
PASS third.selectionEnd is 18
PASS fourth.selectionStart is 11
PASS fourth.selectionEnd is 18
FAIL fourth.selectionStart should be 0. Was 11.
FAIL fourth.selectionEnd should be 0. Was 18.
PASS fifth.selectionStart is 11
PASS fifth.selectionEnd is 18
PASS fifth.selectionStart is 11
PASS fifth.selectionEnd is 18
FAIL sixth.selectionStart should be 11. Was 2.
FAIL sixth.selectionEnd should be 18. Was 2.
FAIL sixth.selectionStart should be 11. Was 2.
FAIL sixth.selectionEnd should be 18. Was 2.
PASS seventh.selectionStart is 11
PASS seventh.selectionEnd is 18
PASS seventh.selectionStart is 11
PASS seventh.selectionEnd is 18
PASS eighth.selectionStart is 11
PASS eighth.selectionEnd is 18
FAIL eighth.selectionStart should be 11. Was 0.
FAIL eighth.selectionEnd should be 18. Was 0.
PASS ninth.selectionStart is 11
PASS ninth.selectionEnd is 18
PASS ninth.selectionStart is 11
PASS ninth.selectionEnd is 18
